// src/component/dashboard/Charts/Widgets/widgets.type.ts
export enum EWidgetName {
  ChatBot = 'Tensurf Brain',
  StatSurf = 'StatSurf',
  TimeSurf = 'TimeSurf',
  LiqSurf = 'LiqSurf',
  PairSurf = 'PairSurf',
  WaveSurf = 'WaveSurf',
  RiskSurf = 'RiskSurf',
  GapSurf = 'GapSurf',
  LevelSurf = 'LevelSurf',
  TrendSurf = 'TrendSurf',
  Idea = 'Idea',
  Request = 'Feature Request',
  BugReport = 'Bug Report'
}

export type TWidgetsName = EWidgetName;